Enhancements to istio.io test framework
- Move Verifier functions to its own file
- Add two Verifiers:
  - ContainsVerifier - checks if the output contains an expected value
  - NotContainsVerifier - checks if the output does not contain a value
- Add SleepCommand: Helper useful between steps to wait config propagation
